fix: refactor to separate firebase code from the hooks
