+++ src/components/common/Loading/__mocks__/index.js
@@ -1,5 +1,5 @@
 import React from 'react';
 
-const Loading = () => <></>;
+const Loading = () => <>Loading</>;
 
 export default Loading;
