+++ src/hooks/useUsersStatistics.js
@@ -1,17 +1,12 @@
 import { useEffect, useMemo } from 'react';
 import { useDispatch, useSelector } from 'react-redux';
-import { collection, doc, getDoc, setDoc, updateDoc } from 'firebase/firestore';
 import { toPairs } from 'lodash';
 
 import { MAX_ATTEMPTS } from 'constants/constants';
-import { USERS_STATISTICS } from 'firebase/collections';
-import firebaseData from 'firebase/firebase';
+import { getUsersStatistics, updateUsersStatistics } from 'firebase/usersStatistics';
 import { setUserStatistics } from 'state/actions/statisticsActions';
 
 import useAuth from './useAuth';
-
-const { firebaseDb } = firebaseData;
-const statisticsRef = collection(firebaseDb, USERS_STATISTICS);
 
 const useUserStatistics = ({ email, name, photo } = {}) => {
   const dispatch = useDispatch();
@@ -30,31 +25,8 @@ const useUserStatistics = ({ email, name, photo } = {}) => {
 
   useEffect(() => {
     const getStatistics = async () => {
-      try {
-        let currentStatistics = {
-          totalGames: 0,
-          totalWins: 0,
-          totalAttempts: Array(MAX_ATTEMPTS + 1).fill(0),
-          currentStreak: 0,
-          lastDatePlayed: '',
-          longestStreak: 0,
-          longestStreakDate: '',
-          attemptedWords: {},
-          timeAverage: 0,
-          minTime: 0,
-          maxTime: 0,
-        };
-        const docRef = doc(firebaseDb, USERS_STATISTICS, selectedUser);
-        const docSnap = await getDoc(docRef);
-        if (docSnap.exists()) {
-          currentStatistics = docSnap.data();
-        } else {
-          await setDoc(doc(statisticsRef, selectedUser), currentStatistics);
-        }
-        await dispatch(setUserStatistics({ statistics: currentStatistics, selectedUser }));
-      } catch (err) {
-        console.error(err);
-      }
+      const { currentStatistics } = await getUsersStatistics(selectedUser);
+      await dispatch(setUserStatistics({ statistics: currentStatistics, selectedUser }));
     };
 
     if (Object.keys(statistics).length === 0) {
@@ -63,7 +35,7 @@ const useUserStatistics = ({ email, name, photo } = {}) => {
   }, [selectedUser, dispatch, statistics]);
 
   const updateStatistics = async newStatistics => {
-    await updateDoc(doc(statisticsRef, selectedUser), newStatistics);
+    await updateUsersStatistics(newStatistics, selectedUser);
     await dispatch(setUserStatistics({ statistics: newStatistics, selectedUser }));
   };
